@@ -5,7 +5,7 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "alpdeskfee": "ng build --prod --output-hashing none && node concate.js",
+    "alpdeskfee": "ng build --configuration production --output-hashing none && node concate.js",
     "alpdeskfeedev": "ng build --output-hashing none && node concatedev.js",
     "test": "ng test",
     "lint": "ng lint",
@@ -13,42 +13,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.5",
-    "@angular/cdk": "^11.2.4",
-    "@angular/common": "~11.2.5",
-    "@angular/compiler": "~11.2.5",
-    "@angular/core": "~11.2.5",
-    "@angular/elements": "^11.2.5",
-    "@angular/forms": "~11.2.5",
-    "@angular/material": "^11.2.4",
-    "@angular/platform-browser": "~11.2.5",
-    "@angular/platform-browser-dynamic": "~11.2.5",
-    "@angular/router": "~11.2.5",
+    "@angular/animations": "~12.2.2",
+    "@angular/cdk": "^12.2.2",
+    "@angular/common": "~12.2.2",
+    "@angular/compiler": "~12.2.2",
+    "@angular/core": "~12.2.2",
+    "@angular/elements": "^12.2.2",
+    "@angular/forms": "~12.2.2",
+    "@angular/material": "^12.2.2",
+    "@angular/platform-browser": "~12.2.2",
+    "@angular/platform-browser-dynamic": "~12.2.2",
+    "@angular/router": "~12.2.2",
     "document-register-element": "^1.7.2",
-    "ngx-build-plus": "^11.0.0",
-    "rxjs": "~6.6.6",
-    "tslib": "^2.1.0",
+    "ngx-build-plus": "^12.2.0",
+    "rxjs": "^6.6.7",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.4",
-    "@angular/cli": "~11.2.4",
-    "@angular/compiler-cli": "~11.2.5",
-    "@types/jasmine": "~3.6.6",
-    "@types/node": "^14.14.34",
-    "codelyzer": "^6.0.1",
+    "@angular-devkit/build-angular": "~12.2.1",
+    "@angular/cli": "~12.2.1",
+    "@angular/compiler-cli": "~12.2.2",
+    "@types/jasmine": "~3.8.1",
+    "@types/node": "^16.3.3",
+    "codelyzer": "^6.0.2",
     "concat": "^1.0.3",
-    "fs-extra": "^9.1.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~6.0.0",
-    "karma": "~6.2.0",
+    "fs-extra": "^10.0.0",
+    "jasmine-core": "~3.8.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.3.4",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
-    "ts-node": "~9.1.1",
+    "ts-node": "~10.1.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.1.5"
+    "typescript": "^4.3.5"
   }
 }
