@@ -1,51 +1,55 @@
-import { Component, Input, OnChanges, OnInit } from '@angular/core';
-import { BaseItemComponent } from '../base-item/base-item.component';
+import {Component, Input, OnChanges, OnInit} from '@angular/core';
+import {BaseItemComponent} from '../base-item/base-item.component';
 
 @Component({
-  selector: 'app-item-custom-module',
-  templateUrl: './item-custom-module.component.html',
-  styleUrls: ['./item-custom-module.component.scss']
+    selector: 'app-item-custom-module',
+    templateUrl: './item-custom-module.component.html',
+    styleUrls: ['./item-custom-module.component.scss']
 })
 export class ItemCustomModuleComponent extends BaseItemComponent implements OnInit, OnChanges {
 
-  @Input() title: string = '';
-  @Input() targetType: string = '';
-  @Input() do: string = '';
-  @Input() iconclass: string = '';
-  @Input() icon: string = '../../../system/themes/flexible/icons/modules.svg';
+    @Input() title = '';
+    @Input() targetType = '';
+    @Input() do = '';
+    @Input() iconclass = '';
+    @Input() icon = '../../../system/themes/flexible/icons/modules.svg';
 
-  @Input() pageEdit: boolean = false;
-  @Input() pageId: number = 0;
+    @Input() pageEdit = false;
+    @Input() pageId = 0;
 
-  iconDefault: string = '../../../system/themes/flexible/icons/modules.svg';
-  iconUrl: string = '';
+    iconDefault = '../../../system/themes/flexible/icons/modules.svg';
+    iconUrl = '';
 
-  ngOnInit() {
-    this.setIcon();
-  }
-
-  ngOnChanges() {
-    this.setIcon();
-  }
-
-  private setIcon() {
-    if (this.icon !== null && this.icon !== '' && this.icon !== undefined) {
-      this.iconUrl = 'url(\'' + this.icon + '\')';
-    } else {
-      this.iconUrl = 'url(\'' + this.iconDefault + '\')';
+    ngOnInit(): void {
+        this.setIcon();
     }
-  }
 
-  click() {
-    this.dispatchEvent({
-      dialog: true,
-      targetType: this.targetType,
-      do: this.do,
-      iconclass: this.iconclass,
-      icon: this.icon,
-      pageEdit: this.pageEdit,
-      pageId: this.pageId
-    });
-  }
+    ngOnChanges(): void {
+        this.setIcon();
+    }
+
+    private setIcon(): void {
+
+        if (this.icon !== null && this.icon !== '' && this.icon !== undefined) {
+            this.iconUrl = 'url(\'' + this.icon + '\')';
+        } else {
+            this.iconUrl = 'url(\'' + this.iconDefault + '\')';
+        }
+
+    }
+
+    click(): void {
+
+        this.dispatchEvent({
+            dialog: true,
+            targetType: this.targetType,
+            do: this.do,
+            iconclass: this.iconclass,
+            icon: this.icon,
+            pageEdit: this.pageEdit,
+            pageId: this.pageId
+        });
+
+    }
 
 }
