@@ -1,10 +1,10 @@
-import { Component } from '@angular/core';
-import { BaseItemComponent } from '../base-item/base-item.component';
+import {Component} from '@angular/core';
+import {BaseItemComponent} from '../base-item/base-item.component';
 
 @Component({
-  selector: 'app-item-move',
-  templateUrl: './item-move.component.html',
-  styleUrls: ['./item-move.component.scss']
+    selector: 'app-item-move',
+    templateUrl: './item-move.component.html',
+    styleUrls: ['./item-move.component.scss']
 })
 export class ItemMoveComponent extends BaseItemComponent {
 
